+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.2.8] - 2025-09-02
+
+### Changed
+- Separator widget for dividing sections in your forms.
+- config/formello.php removed "default_widgets" and added "custom_widgets" to alias your custom widgets.
+
 ## [1.2.7] - 2025-09-01
 
 ### Fixed